Move CatalogDiff modal from views to components
