+++ airbyte-webapp/src/components/connection/CatalogDiffModal/CatalogDiffModal.module.scss
@@ -1,4 +1,4 @@
-@use "../../../scss/variables";
+@use "scss/variables";
 
 .modalContent {
   display: flex;
