+++ airbyte-webapp/src/components/connection/CatalogDiffModal/CatalogDiffModal.test.tsx
