+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffAccordion.module.scss
