+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffAccordionHeader.module.scss
@@ -1,5 +1,5 @@
-@forward "../components/StreamRow.module.scss";
-@forward "../components/DiffSection.module.scss";
+@forward "./StreamRow.module.scss";
+@forward "./DiffSection.module.scss";
 @use "scss/variables";
 
 .row {
