+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffAccordionHeader.tsx
