+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffFieldTable.module.scss
