+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffIconBlock.module.scss
