+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffIconBlock.tsx
