+++ airbyte-webapp/src/components/connection/CatalogDiffModal/DiffSection.module.scss
