+++ airbyte-webapp/src/components/connection/CatalogDiffModal/FieldSection.module.scss
