+++ airbyte-webapp/src/components/connection/CatalogDiffModal/StreamRow.module.scss
