+++ airbyte-webapp/src/components/connection/CatalogDiffModal/index.stories.tsx
