+++ airbyte-webapp/src/components/connection/CatalogDiffModal/index.ts
@@ -1,0 +1,1 @@
+export * from "./CatalogDiffModal";
