+++ airbyte-webapp/src/components/connection/CatalogDiffModal/types.ts
