+++ airbyte-webapp/src/components/connection/CatalogDiffModal/utils.tsx
